requirements.txt
openai>=1.50.0
fastapi>=0.100.0
uvicorn>=0.30.0
streamlit>=1.35.0
arize-phoenix>=4.0.0
pydantic>=2.0.0
redis>=5.0.0
tenacity>=8.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
rich>=13.0.0